requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scipy>=1.10.0
scikit-learn>=1.3.0
japanize-matplotlib>=1.1.3
openpyxl>=3.1.0
xlrd>=2.0.1
chardet>=5.1.0
PyYAML>=6.0
reportlab>=4.0.0
fpdf2>=2.7.0
Pillow>=10.0.0
kaleido>=0.2.1